backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
firebase-admin==6.5.0
python-dotenv==1.0.1
httpx==0.27.0
python-multipart==0.0.9
Werkzeug==3.0.3
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-mock==3.14.0
